# api/openapi.yaml
openapi: 3.1.0
info:
  title: TOT API
  version: 1.0.0-draft
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    JSON API for the Next.js frontend (`frontend-tot`). Draft contract — see
    `docs/frontend/inventory.md` for the page inventory and open decisions.

    Conventions:
    - Base path `/api/v1`; frontend and API share one origin (reverse proxy), so the
      session cookie stays `HttpOnly; SameSite=Lax` and no CORS is needed.
    - JSON bodies and fields are `snake_case`; timestamps are RFC 3339.
    - Errors always use the `Error` schema with a stable machine `code`; user-facing
      text is localized on the frontend by `code`, `message` is for logs only.
    - Unauthenticated requests get `401`, never a redirect; malformed JSON gets `400 malformed_json`.
    - Mutating requests (POST/PUT/PATCH/DELETE) must come from an allowed origin
      (`Origin` / `Sec-Fetch-Site`), otherwise `403 csrf_rejected`.
    - Request bodies are `application/json` unless stated otherwise, otherwise `415 unsupported_media_type`.
    - HTML fields (`*_html`) are rendered and sanitized server-side.
    - Admin endpoints need role `author` or `admin`; admin-only ones say so. A course
      the author cannot access answers `404`; a missing permission on an accessible
      course answers `403 forbidden` (see `CourseAccess`).
servers:
  - url: /api/v1
...
security:
  - session: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Login
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Signed in; sets the `session` cookie.
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401': { $ref: '#/components/responses/InvalidCredentials' }
        '415': { $ref: '#/components/responses/UnsupportedMedia' }
        '403':
          description: '`account_blocked`.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/register:
    post:
      tags: [auth]
      summary: Register
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
                email: { type: string, maxLength: 254 }
                password: { type: string, minLength: 6, maxLength: 72 }
      responses:
        '201':
          description: Account created and signed in.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '403':
          description: '`registration_closed`.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: '`email_taken`.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '415': { $ref: '#/components/responses/UnsupportedMedia' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /me:
    get:
      tags: [me]
      summary: Get me
      operationId: getMe
      responses:
        '200':
          description: Current user (replaces the `gl_role` / `gl_user` hint cookies).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/profile:
    get:
      tags: [me]
      summary: Get profile
      operationId: getProfile
      responses:
        '200':
          description: Profile page data.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/dashboard:
    get:
      tags: [me]
      summary: Get dashboard
      operationId: getDashboard
      description: |
        Progress dashboard. Course cards come from `GET /catalog`; the heatmap grid
        and month labels are built on the frontend from `activity`.
      responses:
        '200':
          description: Dashboard data.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dashboard' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ───────────────────────────── public ─────────────────────────────
...
  /catalog:
    get:
      tags: [catalog]
      summary: Get catalog
      operationId: getCatalog
      description: |
        Published specializations with their published courses and the user's progress.
        Feeds the courses page, roadmap, dashboard course grid and trainers list.
        Gym courses (`track=gym`) are returned separately in `trainers`.
      responses:
        '200':
          description: Catalog.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Catalog' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /specializations/{specSlug}:
    parameters:
      - $ref: '#/components/parameters/SpecSlug'
    get:
      tags: [catalog]
      summary: Get specialization
      operationId: getSpecialization
      description: Draft specializations are `404` except for admins.
      responses:
        '200':
          description: Specialization with its courses.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SpecializationWithCourses' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /specializations/{specSlug}/cover:
    parameters:
      - $ref: '#/components/parameters/SpecSlug'
    get:
      tags: [catalog]
      summary: Get specialization cover
      operationId: getSpecializationCover
      security: []
      responses:
        '200':
          description: Uploaded image or generated SVG.
          content:
            image/*: { schema: { type: string, format: binary } }
            image/svg+xml: { schema: { type: string } }
        '302': { description: Redirect to an external cover URL. }
        '404': { $ref: '#/components/responses/NotFound' }
  /courses/{courseSlug}:
    parameters:
      - $ref: '#/components/parameters/CourseSlug'
    get:
      tags: [catalog]
      summary: Get course
      operationId: getCourse
      description: Course page with published lessons. Draft courses are `404` except for admins.
      responses:
        '200':
          description: Course with its chapter list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CourseDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /courses/{courseSlug}/cover:
    parameters:
      - $ref: '#/components/parameters/CourseSlug'
    get:
      tags: [catalog]
      summary: Get course cover
      operationId: getCourseCover
      security: []
      responses:
        '200':
          description: Uploaded image or generated SVG.
          content:
            image/*: { schema: { type: string, format: binary } }
            image/svg+xml: { schema: { type: string } }
        '302': { description: Redirect to an external cover URL. }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────────────────────── lessons ─────────────────────────────
  /courses/{courseSlug}/lessons/{lessonSlug}:
    parameters:
      - $ref: '#/components/parameters/CourseSlug'
      - $ref: '#/components/parameters/LessonSlug'
    get:
      tags: [lessons]
      summary: Get lesson
      operationId: getLesson
      description: |
        Theory / quiz / sql / sim lesson. Pure read — no progress side effects
        (call `POST /lessons/{lessonId}/visit`). For `kind=lab` use `GET /lessons/{lessonId}/lab`.
        Quiz questions come without correct answers. Draft lessons and courses are `404` except for admins.
      responses:
        '200':
          description: Lesson.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LessonDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /lessons/{lessonId}/visit:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    post:
      tags: [lessons]
      summary: Visit lesson
      operationId: visitLesson
      description: Marks the lesson `in_progress` if the user has no progress yet. Idempotent.
      responses:
        '200':
          description: Current progress.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LessonProgress' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /lessons/{lessonId}/complete:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    post:
      tags: [lessons]
      summary: Complete lesson
      operationId: completeLesson
      description: Manual completion; allowed only for `theory`, `sql` and `sim` lessons.
      responses:
        '200':
          description: Updated progress.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LessonProgress' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`completion_derived` — quiz/lab lessons complete via their checks.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /lessons/{lessonId}/notes:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    put:
      tags: [lessons]
      summary: Save lesson notes
      operationId: saveLessonNotes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [notes]
              properties:
                notes: { type: string, maxLength: 20000 }
      responses:
        '204': { description: Saved. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }

  # ───────────────────────────── quiz ─────────────────────────────
  /lessons/{lessonId}/quiz/answers:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    post:
      tags: [quiz]
      summary: Answer quiz question
      operationId: answerQuizQuestion
      description: |
        Answers one question. The first answer per question is stored and locked
        until the attempt is reset; repeating returns the stored result with `locked=true`.
        Reveals the correct option only after an answer is stored.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuizAnswerInput' }
      responses:
        '200':
          description: Answer result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuizAnswerResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      tags: [quiz]
      summary: Reset quiz attempt
      operationId: resetQuizAttempt
      description: Clears stored answers of the current attempt ("Заново"). The saved score is kept.
      responses:
        '204': { description: Attempt reset. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /lessons/{lessonId}/quiz/submit:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    post:
      tags: [quiz]
      summary: Submit quiz
      operationId: submitQuiz
      description: |
        Scores the attempt, records it in the attempt history and saves `quiz_score`/`quiz_total`. Questions answered via
        `POST .../answers` keep the stored answer; `answers` fills (and stores) the rest.
        Unanswered questions count as wrong. `404` when the lesson has no quiz.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: array
                  items: { $ref: '#/components/schemas/QuizAnswerInput' }
      responses:
        '200':
          description: Quiz result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/QuizResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }

  # ───────────────────────────── labs ─────────────────────────────
  /lessons/{lessonId}/lab:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    get:
      tags: [labs]
      summary: Get lab
      operationId: getLab
      description: Lab page (tasks, theory, quiz, sandbox state). Never exposes solutions or check scripts. `404` when the lesson has no tasks.
      responses:
        '200':
          description: Lab.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lab' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /lessons/{lessonId}/lab/session:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    get:
      tags: [labs]
      summary: Get lab session
      operationId: getLabSession
      description: Sandbox state; used to re-attach the terminal after reload.
      responses:
        '200':
          description: Session state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SandboxSession' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [labs]
      summary: Stop lab session
      operationId: stopLabSession
      description: Tears the sandbox down (reset environment). Progress is kept.
      responses:
        '204': { description: Teardown started. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /lessons/{lessonId}/lab/retry:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    post:
      tags: [labs]
      summary: Retry lab
      operationId: retryLab
      description: Deletes the user's submissions and current quiz answers for the lab, resets progress (notes and attempt history kept) and tears the sandbox down.
      responses:
        '204': { description: Lab reset. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /lessons/{lessonId}/lab/terminal:
    parameters:
      - $ref: '#/components/parameters/LessonId'
      - $ref: '#/components/parameters/TermCols'
      - $ref: '#/components/parameters/TermRows'
    get:
      tags: [labs]
      summary: Open lab terminal
      operationId: openLabTerminal
      description: |
        WebSocket upgrade — see `x-websocket` for the frame protocol.
        Opening a terminal tears down any other sandbox of the same user.
      x-websocket: terminal  # see components.x-websocket-protocols
      responses:
        '101': { description: Switching protocols. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/SandboxDisabled' }
  /lessons/{lessonId}/lab/git-graph:
    parameters:
      - $ref: '#/components/parameters/LessonId'
    get:
      tags: [labs]
      summary: Get lab git graph
      operationId: getLabGitGraph
      description: Parsed `git log --all` of `/root/project` (latest 60 commits, newest first). Empty list if no repo or no running sandbox; never boots one.
      responses:
        '200':
          description: Commits.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GitGraph' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/SandboxDisabled' }
  /lessons/{lessonId}/lab/fs/entries:
    parameters:
      - $ref: '#/components/parameters/LessonId'
      - $ref: '#/components/parameters/SandboxPath'
    get:
      tags: [labs]
      summary: List lab files
      operationId: listLabFiles
      description: One directory level inside `/root`, directories first.
      responses:
        '200':
          description: Directory listing.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LabDirListing' }
        '400': { $ref: '#/components/responses/PathOutsideJail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/SandboxNotRunning' }
        '502': { $ref: '#/components/responses/SandboxError' }
        '503': { $ref: '#/components/responses/SandboxDisabled' }
  /lessons/{lessonId}/lab/fs/content:
    parameters:
      - $ref: '#/components/parameters/LessonId'
      - $ref: '#/components/parameters/SandboxPath'
    get:
      tags: [labs]
      summary: Read lab file
      operationId: readLabFile
      description: Raw file bytes, up to 2 MiB.
      responses:
        '200':
          description: File content.
          content:
            application/octet-stream: { schema: { type: string, format: binary } }
        '400': { $ref: '#/components/responses/PathOutsideJail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413':
          description: '`file_too_large`.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409': { $ref: '#/components/responses/SandboxNotRunning' }
        '502': { $ref: '#/components/responses/SandboxError' }
        '503': { $ref: '#/components/responses/SandboxDisabled' }
    put:
      tags: [labs]
      summary: Write lab file
      operationId: writeLabFile
      description: Writes raw bytes (max 2 MiB), creating parent directories.
      requestBody:
        required: true
        content:
          application/octet-stream: { schema: { type: string, format: binary } }
      responses:
        '204': { description: Written. }
        '400': { $ref: '#/components/responses/PathOutsideJail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '413':
          description: '`file_too_large`.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409': { $ref: '#/components/responses/SandboxNotRunning' }
        '502': { $ref: '#/components/responses/SandboxError' }
        '503': { $ref: '#/components/responses/SandboxDisabled' }
...
  /tasks/{taskId}/check:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags: [labs]
      summary: Check task
      operationId: checkTask
      description: Runs the task's check script in the sandbox; exit 0 = passed. Saves the submission.
      responses:
        '200':
          description: Check result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskCheckResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`task_not_auto_checked` — use `POST /tasks/{taskId}/done`.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '502': { $ref: '#/components/responses/SandboxError' }
        '503': { $ref: '#/components/responses/SandboxDisabled' }
  /tasks/{taskId}/done:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    post:
      tags: [labs]
      summary: Mark task done
      operationId: markTaskDone
      description: Manual completion for tasks without a check script or test cases.
      responses:
        '200':
          description: Updated task state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskCheckResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`task_auto_checked` — the task must pass its check.'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /playground/run:
    post:
      tags: [trainers]
      summary: Run playground
      operationId: runPlayground
      description: Go playground. Compile/validation errors are returned in `errors` with 200.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RunRequest' }
      responses:
        '200':
          description: Run result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RunResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '502': { $ref: '#/components/responses/SandboxError' }
  /git-trainer/session:
    get:
      tags: [trainers]
      summary: Get git trainer session
      operationId: getGitTrainerSession
      responses:
        '200':
          description: Session state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SandboxSession' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [trainers]
      summary: Reset git trainer
      operationId: resetGitTrainer
      description: Tears the git trainer sandbox down; the next terminal gets a fresh repo.
      responses:
        '204': { description: Reset. }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: session
      description: HttpOnly session cookie set by `POST /auth/login`.
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
...